debug-connectivity: ensure kafkacat is successful and 2+ endpoints are tested
When kafkacat fails, it emits on STDOUT instead of STDERR but exits
non-zero.  Ensure that these errors are caught and clear to the user.

The expectation is that bootstrap + brokers are all tested, if for some
reason, no brokers are found, this should expectation should be
explained to the user.