Add bootstrap prefix as Security Group suffix
We might want to use this module in the same AWS account,
but we are creating one SG with a fixed name.

Here we use the privatelink endpoint prefix (eg. lkc-abcde-vwxyz)
as suffix for the SG name, and also add a random id,
so it is unique for each cluster and easy to identify.